fix: some usages of MySK logo not renamed
