+++ components/home/sake/SakeBackground.tsx
@@ -1,6 +1,6 @@
 import BlobsBlurringGroup from "@/components/common/BlobsBlurringGroup";
 import DecorativeBlob from "@/components/common/DecorativeBlob";
-import MySKLogo from "@/public/images/brand/mysk-light.svg";
+import MySKLogo from "@/public/icons/petals-light.svg";
 import cn from "@/utils/helpers/cn";
 import { StylableFC } from "@/utils/types/common";
 import Image from "next/image";
